added in a label in the create ingredients window that says if the ingredient was added succesfully or not. also it clears the fields when entered....the back button works too.
In the create Recipe window I wrote a bunch of psudo code that is commented out and fleshed out its basic funcionality
while at the same time cleaning up the interface and added in a steps widget to type in the steps for the recipe.

In the create meal window that ui is a little bit bugged when ran, some of the labels get pushed up to the upper left corner
for w/e reason, a minor bug that doesnt really matter at the moment.  Also in that window I wrote psudo code to populate the combo box with the list of recipes.....